style: finish clippy/dead-code cleanup — zero warnings across all targets (Task 183)
- Remove dead test helpers (node_insert, first_clause)
- Replace PI-approximating test literals (3.14 -> 2.5) to satisfy approx_constant
- Justified allows for intentional structural/API naming (module_inception on
  graph::graph, enum_variant_names on BuilderError, large_enum_variant on
  ServerStreamInner) and test-only field-reassign-with-default
- cargo clippy --all-targets: 0 warnings; cargo check --all-targets: 0 errors

+++ src/storage/manager.rs
@@ -613,7 +613,7 @@ mod tests {
         handle
             .read_at(
                 &mut bitmap_buf,
-                FIRST_BITMAP_PAGE_ID as u64 * PAGE_SIZE as u64,
+                FIRST_BITMAP_PAGE_ID * PAGE_SIZE as u64,
             )
             .unwrap();
         drop(handle);
